Improvements to Android printf and main wrappers.
This change discards the previous preprocessor-based redefining of
the stdio functions for a cleaner approach. It also adds examples,
replaces exposed globals with accessors, adds unit tests and gtest
integration, removes the old ostream impl in favor of simply making
iostreams work, and other cleanup.

The main wrapper is cleaned up as well, and split into a separate
library for conditional linking, allowing users to opt-in to either
or both of the main wrapper and stdio wrappers.

Tested: Ran example and tests.
Bug: 13915748
Change-Id: I692dba0287491496ffac16ff23cb478b962d65f1