Ignore preview SDKs reported by "android list target --compact".
"android list target --compact" currently reports the "L" preview
SDK as "android-L".  Since there doesn't appear to be a clean
mapping from preview SDK to a SDK version level, this ignores
preview SDKs when selecting an appropriate SDK to build against.

Tested:
Verified that it's possible to build with ADT 20140702.

Bug: 17691796
Change-Id: I346d22d9655ffd795a2eaa6278d14d8badba2dc3