Excluded dependencies from examples of build_all_android
googletest includes an incomplete Android project build causes
build_all_android to fail when recursively building all projects
in the tree.  This changes the documentation to exclude the
dependencies/ directory from the set of projects that are built
in examples of build_all_android.

Tested:
Built the docs, verified they look correct.
Ran build_all_android in a tree with the dependencies/ directory
present.

Change-Id: Iab0a8801f53732a7c54e95aec000e31be8516849