Create IndexAllocator.
The IndexAllocator lets you allocate and free indices. It tries to
keep the indices as low as possible. It provides a mechanism to
backfill all holes and make the allocated indices contiguous.

Useful for data-oriented design, where contiguous memory is
essential for performance.

Tested: Added new set of tests. Passes on Linus. Having trouble
deploying on Android, but builds.
Change-Id: Id492cd88d97bbba882756c978b4b94ed593b8870