Add file utilities. Add fplutil library.
We didn't build a library before (except in android) because
the allocator code was all in header files. Now that we've added
a proper .cpp file, we need to build a lib.

Also, changed all CMAKE_SOURCE_DIR into CMAKE_CURRENT_SOURCE_DIR.
CMAKE_SOURCE_DIR refers to the location of the calling
CMakeLists.txt. We want the location of *our* CMakeLists.txt.
This was causing a failure when we did add_subdirectory(fplutil)
from fplbase.

Tested: Links agains mesh_pipeline executable in fplbase, on Linux.
Change-Id: Iab3f74c37990a8f34ee388438136309f143e050d