Added a testing mode to push_package.py which creates a tree of symlinks.
It's now possible to run...

./push_package.py -S ~/symlinks_dir -b master

to create a set of symlinks to the project and its dependencies in order to
verify the package works in the published directory structure.

In addition, this enables symlink traversal in build_all_android and fixes a
build failure when attempting to build the index allocator test.

Tested:
Published fplutil to a local directory and verified everything builds.

Bug: 21620188
Change-Id: If66b2b422d7ca11f956d415ec8dc07d4f643749e