Redo Defrag algorithm in IndexAllocator.
New algorithm is simpler and more robust, particularly during
callback functions. It also has significantly better worst-case
performance.

Count now must be the same type as Index. This is necessary since
there coule be a really long strand of consecutive, freed indices.
It also just makes life simpler since there are fewer template
parameters.

Count is signed and so Index is now also forced to be signed.
That's purely an internal algorithm requirement (we could work
around it, but it's easier to debug with this requirement).
This is a minor restriction that I doubt will ever be a problem.

Tested: Added new unit tests and expanded unit tests to check
compinations of algorithm internals.
Change-Id: Idb3bf5bc9bfa3b4d41717fb1bac0feb2509adf96