Sample running support tools and functionality.
1. Support for multiple activities in a manifest.
2. Toggleable SDK version validation.
3. User specified 'end-condition' regexes for APK runs.

Each of these is useful for my usecase of trying to run Android samples.
Many of the samples have multiple activities and store their SDK in
build.gradle files. I would also like to run an APK and watch it for a
crash/ANR, but still terminate the process without a keyboard interrupt.

Tested: Added & ran unit tests. Ran on several samples checking for ANR,
crashes, and success.

Change-Id: I0e093752a78535644ce20f3a5cd29d24ac45653c