Including last line of logcat and adb emulator fix
Some emulators only report the device serial and type in adb devices -l.
This includes a fix where some emulators would not result in a correct
instantiation of AdbDevice.

I need the last line in the logcat to correctly report what the problem
is in my sample runner, and the module currently excludes it.

Also added __init__.py to the root to allow external python modules to
refer to fplutil.buildutil.

Tested: android_test.py and ran with custom runner.

Change-Id: I597d843f749622a06adf3fe2d8714d978121498a