Set the shared docs directory by default.
generate_docs.py to now sets SHARED_DOCS_PATH to the directory
containing the generate_docs.py script if it isn't already set.
I needed to import the generate_docs module *again* to get the path
of it since the __file__ property of a module seems to point at the
entry point (e.g if A imports B, __file__ referenced in B will point to A).

Tested:
Removed Ib4ce89 from generate_docs.py in MathFu and successfully rebuilt
the docs including the common header and footer.

Bug: 22506506
Change-Id: Id5c27ca01b84fc014b88a2c17c545be884fce1ed