Clean intermediate files from NDK builds when the NDK location changes
NDK makefiles contain absolute paths to the NDK which means that if
the NDK is moved, builds will fail in strange ways.  This keeps track
of which NDK was used to build each project and forces a rebuild if
the location of the NDK changes.

Tested:
1. Built fplutil on OS X for Android.
2. Moved the NDK
3. Rebuilt fplutil and verified everything was cleaned and rebuilt.
4. Built again and verified nothing was cleaned and rebuilt.
5. Moved the NDK.
6. Verified everything is cleaned and rebuilt.

Unit tests for this functionality is missing but the existing unit
tests continue to pass.

Bug: 23160373
Change-Id: I4cb8b228936560f47b03cb5d3a9378fa211d9795