Modified intrusive list to work around MSVC bug.
Visual Studio fails to compile in the case where an intrusive_list_node
is contained on the structure that is holding the intrusive_list. This
might come up when, for example, you try to construct a tree where each
object has a list child objects of the same type.

This change makes the node_offset into a value that is determined at run
time instead of compile time. This adds a few bytes to the list itself
and changes how it gets initialized, but it is otherwise mostly the same
to the user.

Tested on Linux, all unit tests pass.

Change-Id: I7754cf7a8ef68144bda702c76bbf57a1017bf662