Fix storing NDK build location between builds.
We want to delete the output directory if the NDK build location has
changed. The problem was the file that stored the current NDK build
location was deleted as well, so every build was considered to have a
new NDK build location, hence forcing a rebuild of everything.

Tested by repeated builds on Android.

Change-Id: I5f01f3059b066633b4d90c1a2468362e070b004f