Fix Visual Studio 2012 build.
There was a bug in the directory creation function.

Also, the intrusive list assignment operator is
being linked in. It's not clear if it's actually being
called or not. Added asserts so we can see, once the
Windows build boots correctly.

Tested: Builds on Visual Studio 2012. Runs with
mesh_pipeline and asset_pipeline. Runs on Linux.

Change-Id: I873d0b5fac902abb35d697ead818747674a73645