Adding helpful error message to generate_docs.py.
If you do not have 'doxygen' installed, the error information provided
by 'generate_docs.py' is very uninformative. This occurs because the
code attempts to print out 'e.status' (where 'e' is the error). However,
the error thrown is of type 'CalledProcessError', which does not contain
a 'status' attribute.

As a result, I removed the reference to 'e.status', and simply print a
string literal instead.  This instructs the user that the error occurred
because they need to install 'doxygen'.

Bug: b/24601209
Change-Id: I54d2a3cc8cbce1cae583c8bc075e1ce7800635ab
Tested: Running 'generate_docs.py' on both OS X and Ubuntu.