Changed symlink staging to a mirror.
Creating a tree of symlinks to test out distribution source builds
causes a lot of pain with both CMake and ndk-build.  In both cases
we end up with many paths in our builds that look like this...

  a/b/../../c/d

If a/b in the above example is a symlink both CMake and ndk-build/make
will walk along each node in the directory tree and go "up" and out of
the directory referenced by the symlink a/b.

So to keep things simple, this commit changes push_package.py to
support a "mirror" option where all files from a package and
its dependencies are copied into a mirror directory.

Tested:
Verified this works with the motive package.

Bug: 22510457
Change-Id: I2b19416adba8bc2568a2cbd67a2d703e842c77ce