Fixed the python package path errors in docs.
The doxygen generated docs were creating wonky package paths
because the staging area (root folder) contained an __init__.py.

To fix this, the push_package.py now creates an base folder with
the project name to ensure the __init__.py in the root will be
in the correct path for doxygen to generate appropriate packages.

I also did some miscellaneous cleanup in the docs to make sure
the pydocs get parsed correctly in python.

Bug: b/25948810

Tested: Tested using push_package.py on Ubuntu with a temporary
staging area and a given staging area with the `-s` flag. All
docs were viewed on Chrome.

Change-Id: I02ee22d6ce923041b4f722bb4e142a828adab89f