Adding support to handle iOS SDK versions in docs.
Added a JSON file the serves as a reference for
tested/supported iOS SDK versions. Currently, this is only
used in the build docs for FPLBase, but it can eventually
be used in other libraries as well.

Added a JS file that can be included in any docs to allow
easy templating of the build documentation by parsing the
JSON file and modifying the resulting page's HTML
appropriately.

Change-Id: Id68014bee3c01d59027dfccee1a067b4d0dd156b
Tested: Tested with generate_docs.py (doxygen) on Ubuntu
and viewed on Chrome.